修改路由api不需要登录访问的接口

+++ app/Http/Controllers/Api/UserAddressController.php
@@ -155,6 +155,9 @@ class UserAddressController extends Controller
      */
     public function destroy(UserAddress $address)
     {
+        if ($address->user_id != Auth::guard('api')->id()) {
+            return $this->failed('禁止访问', 403);
+        }
         $result = null;
         try {
             $result = $address->delete();
